feat: harmonize usage of the Dialog component
